Add files via upload
This is a system that analyses user login behaviour, providing real-time insights and visualizations in the form of interactive charts and graphs. The system will include secure authentication using JWT and Bcrypt, report user login time, and ensure that anomalies are detected if there are any deviations from regular patterns. The system will send out email notifications upon detection of anomalies and detailed daily reports. The backend was developed using Node.js. Data was hosted on MongoDB, and Web Sockets were added for real-time updates. The analysis was visualized using Chart.js. It added different graphs for regular logins and new users, which will help the administrators understand the trend behind login activities and solve the cause of the problems of login efficiency.